refactor: move ptrace tracer into ptrace mod
