+++ src/ptrace.rs
@@ -1,9 +1,11 @@
 mod engine;
 mod guards;
+mod inspect;
 mod syscall;
+mod tracer;
 mod waitpid;
-pub mod inspect;
 
 pub use engine::RecursivePtraceEngine;
 pub use guards::*;
+pub use tracer::*;
 pub use waitpid::*;
